Hide path

+++ .zshrc
@@ -123,12 +123,19 @@ alias de='docker compose exec $(docker compose ps --services | grep _api) entryp
 function da() {
     docker attach $(docker compose ps | grep _api | awk '{print $1}')
 }
-alias bc='docker compose -f /Users/m-ohba/ghq/github.com/smartbank-inc/Bocal/docker-compose.yml'
-alias bcr='bc run --rm $(pwd | xargs basename | sed -e "s/[a-z][0-9]\{2\}-//g")'
-alias be='bc exec $(pwd | xargs basename | sed -e "s/[a-z][0-9]\{2\}-//g") entrypoint.sh'
-function ba() {
-    docker attach $(bc ps | grep _api | awk '{print $1}')
-}
+
+# Run docker-compose command with specific file path.
+# The path may be a kind of secret information, so it's retrieved via macOS keychain
+# ref https://qiita.com/kroyagis/items/66ca139a4c41b710a53c
+export CUSTOM_DOCKER_COMPOSE_PATH=$(security find-generic-password -s "docker-compose-path" -w)
+if [[ -f "$CUSTOM_DOCKER_COMPOSE_PATH" ]]; then
+  alias bc='docker compose -f $CUSTOM_DOCKER_COMPOSE_PATH'
+  alias bcr='bc run --rm $(pwd | xargs basename | sed -e "s/[a-z][0-9]\{2\}-//g")'
+  alias be='bc exec $(pwd | xargs basename | sed -e "s/[a-z][0-9]\{2\}-//g") entrypoint.sh'
+  function ba() {
+      docker attach $(bc ps | grep _api | awk '{print $1}')
+  }
+fi
 
 # key binding
 bindkey '^a' beginning-of-line
